chore: split errors into dirs
