--- crates/committer/src/patricia_merkle_tree/errors.rs
@@ -1,39 +1,0 @@
-use thiserror::Error;
-
-use crate::patricia_merkle_tree::node_data::leaf::LeafDataTrait;
-use crate::patricia_merkle_tree::types::NodeIndex;
-use crate::storage::errors::{DeserializationError, StorageError};
-
-use crate::patricia_merkle_tree::filled_tree::node::FilledNode;
-
-#[allow(dead_code)]
-#[derive(Debug, Error)]
-pub(crate) enum OriginalSkeletonTreeError {
-    #[error(
-        "Failed to deserialize the storage value: {0:?} while building the original skeleton tree."
-    )]
-    Deserialization(#[from] DeserializationError),
-    #[error(
-        "Unable to read from storage the storage key: {0:?} while building the \
-         original skeleton tree."
-    )]
-    StorageRead(#[from] StorageError),
-}
-
-#[derive(Debug)]
-#[allow(dead_code)]
-pub(crate) enum UpdatedSkeletonTreeError<L: LeafDataTrait> {
-    MissingNode(NodeIndex),
-    DoubleUpdate {
-        index: NodeIndex,
-        existing_value: Box<FilledNode<L>>,
-    },
-    PoisonedLock(String),
-    NonDroppedPointer(String),
-}
-
-#[derive(thiserror::Error, Debug, derive_more::Display)]
-pub(crate) enum FilledTreeError {
-    MissingRoot,
-    SerializeError(#[from] serde_json::Error),
-}
