+++ crates/committer/src/patricia_merkle_tree/filled_tree/errors.rs
@@ -1,0 +1,5 @@
+#[derive(thiserror::Error, Debug, derive_more::Display)]
+pub(crate) enum FilledTreeError {
+    MissingRoot,
+    SerializeError(#[from] serde_json::Error),
+}
