+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree.rs
@@ -1,3 +1,4 @@
+pub mod errors;
 pub mod node;
 pub mod original_skeleton_calc;
 pub mod tree;
