+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree.rs
@@ -1,3 +1,4 @@
+pub mod errors;
 pub mod hash_function;
 pub mod node;
 pub mod tree;
