+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree/errors.rs
@@ -1,0 +1,16 @@
+use crate::patricia_merkle_tree::node_data::leaf::LeafDataTrait;
+use crate::patricia_merkle_tree::types::NodeIndex;
+
+use crate::patricia_merkle_tree::filled_tree::node::FilledNode;
+
+#[derive(Debug)]
+#[allow(dead_code)]
+pub(crate) enum UpdatedSkeletonTreeError<L: LeafDataTrait> {
+    MissingNode(NodeIndex),
+    DoubleUpdate {
+        index: NodeIndex,
+        existing_value: Box<FilledNode<L>>,
+    },
+    PoisonedLock(String),
+    NonDroppedPointer(String),
+}
